Refactoring to create one big TCP server effect.
Having lots of different effects just to set up the tcp server seemed
unnecessary. So this include a big effect that just handles the
server, on connection it spawns a thread and runs it in our effect.

This isn't quite perfect yet, it is theoretically possible to run the
client thread without having a connected socket. I need to work out
how to make this so it doesn't compile if you try to run without.